backend/requirements.txt
fastapi
uvicorn
pydantic
Authlib
passlib[bcrypt]
python-jose[cryptography]
email-validator
